requirements.txt
# Core dependencies
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0

# Data providers
polygon-api-client==1.12.4
yfinance==0.2.36

# Scheduling
schedule==1.2.0
pytz==2023.3

# Alerts
python-telegram-bot==20.7
discord-webhook==1.3.0

# Dashboard
streamlit==1.29.0
plotly==5.18.0

# Logging and utilities
loguru==0.7.2
pydantic==2.5.2
pyyaml==6.0.1
